Allow STEP as an unsigned u64 by doing all the math as i128
The STEP value is unsigned, and so you can have

```fish
set -l max 9223372036854775807
set -l min -9223372036854775807
set -l diff_max 18446744073709551614

random -- $min $diff_max $max
```

which, in C++, will work out and randomply pick either the minimum long long or the maximum long long. (same for just having a range that's larger than the maximum long long in general, like going from the minimum to 0)

It does that by doing a whole bunch of gnarly math that I simply
haven't been able to reproduce. So instead we say fudge it, and we do
the readable thing of

1. Pick the number of possibilities
2. Generate a number in that range
3. Add it, multiplied by the step value, to the start

and do all this in 128-bits, so it'll all work out, and then check for
overflow *after* - if it's too large for 64-bits (signed or unsigned!), we error out.