Standardize (un)escape_string_xxx() functions
These are intended to have fixed, reproducible behavior but they were actually
locale-sensitive and did not necessarily encode non-ASCII input as hex-encoded
UTF-8 sequences (and instead used whatever encoding wctombb() returned).

Additionally, the unescape functions had code that could not be reached (such as
checking if c == '_' when c was not ASCII) and duplicate that wasn't needed
(such as checking for '\0' before checking if a char is a hex digit). The new
unescape functions might also be stricter in rejecting invalid input.

The escape functions no longer convert the entirety of the wide input into a
narrow string, thereby avoiding unnecessary allocations. It is similarly
possible to decode directly into a WString result in the unescape functions, but
it would increase branchiness and complexity of the code in order to handle
sequences of 2, 4, or 8 hex characters per wide char instead of just appending
all UTF-8 bytes and then widening the string at the end of the operation.