Use interior mutability for most Job and Process operations
We only use `&mut Job` from the creation of a new `Job` until we put it in an
`Arc` and hand it over to the parser. From that point on, all `Job` access must
be read-only.

Aside from not needing an `RwLock` because a Job is never sent to another
thread and is only accessed from the same thread it was created, I believe the
old code was prone to deadlocks since there were too many read borrows flying
around at once for us to reliably say that no one will try to get a simultaneous
write lock.

The change necessitated the same treatment for `Process`: aside from when it is
first created, we no longer use `&mut Process` for anything. All methods that
can be called during the lifetime of a Process now use interior mutability.

Note that with these changes we need to make sure that Job, JobRef, Process,
etc. are all !Send and !Sync. Our interior mutability uses atomics but it
doesn't use locks when updating multiple fields, meaning outside observers from
another thread would see torn writes to the structs as a whole.

With this change, the usages of `Parser::jobs_mut()` have been whittled down to
just two (those that add a new job and those that remove an existing job) - all
other access can now happen via `&Job` read-only references.