wildcard.cpp: Skip file_id_t on a final segment ending in a "/"
This makes it so expand_intermediate_segment knows about the case
where it's last, only followed by a "/".

When it is, it can do without the file_id for finding links (we don't
resolve the files we get here), which allows us to remove a stat()
call.

This speeds up the case of `*/` or `.../...*.../` by quite a bit.

If that last component was a directory with 1000 subdirectories we
could skip 1000 stat calls!